backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic[email]==2.12.5
email-validator==2.3.0
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
yfinance>=0.2.44,<2.0
python-multipart==0.0.5